refactor: dto directory & add EventImage Entity
