+++ src/main/java/com/drinkhere/drinklystore/application/presentation/event/EventController.java
@@ -1,0 +1,12 @@
+package com.drinkhere.drinklystore.application.presentation.event;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/event/m")
+@RequiredArgsConstructor
+public class EventController {
+
+}
