+++ src/main/java/com/drinkhere/drinklystore/application/service/Impl/event/CreateEventUseCase.java
@@ -1,0 +1,13 @@
+package com.drinkhere.drinklystore.application.service.Impl.event;
+
+import com.drinkhere.drinklystore.common.annotation.ApplicationService;
+import com.drinkhere.drinklystore.domain.service.event.EventCommandService;
+import lombok.RequiredArgsConstructor;
+
+@ApplicationService
+@RequiredArgsConstructor
+public class CreateEventUseCase {
+    private final EventCommandService eventCommandService;
+
+    public
+}
