+++ src/main/java/com/drinkhere/drinklystore/application/service/Impl/store/GetOwnerMainPageUseCase.java
@@ -1,8 +1,8 @@
 package com.drinkhere.drinklystore.application.service.Impl.store;
 
 import com.drinkhere.drinklystore.common.annotation.ApplicationService;
-import com.drinkhere.drinklystore.domain.dto.response.GetFreeDrinkHistoryResponse;
-import com.drinkhere.drinklystore.domain.dto.response.GetOwnerMainPageResponse;
+import com.drinkhere.drinklystore.domain.dto.freedrink.response.GetFreeDrinkHistoryResponse;
+import com.drinkhere.drinklystore.domain.dto.store.response.GetOwnerMainPageResponse;
 import com.drinkhere.drinklystore.domain.entity.store.FreeDrinkHistory;
 import com.drinkhere.drinklystore.domain.entity.store.Store;
 import com.drinkhere.drinklystore.domain.service.freedrinkhistory.FreeDrinkHistoryQueryService;
