+++ src/main/java/com/drinkhere/drinklystore/application/service/Impl/store/MenuUpdateImagesUseCaseImpl.java
@@ -2,9 +2,9 @@ package com.drinkhere.drinklystore.application.service.Impl.store;
 
 import com.drinkhere.drinklystore.application.service.UpdateImagesUseCase;
 import com.drinkhere.drinklystore.common.annotation.ApplicationService;
-import com.drinkhere.drinklystore.domain.dto.request.StoreImageUpdateRequest;
-import com.drinkhere.drinklystore.domain.dto.response.ImageInfoResponse;
-import com.drinkhere.drinklystore.domain.dto.response.StoreResponse;
+import com.drinkhere.drinklystore.domain.dto.store.request.StoreImageUpdateRequest;
+import com.drinkhere.drinklystore.domain.dto.store.response.ImageInfoResponse;
+import com.drinkhere.drinklystore.domain.dto.store.response.StoreResponse;
 import com.drinkhere.drinklystore.domain.entity.store.Store;
 import com.drinkhere.drinklystore.domain.enums.StoreImageType;
 import com.drinkhere.drinklystore.domain.service.store.StoreCommandService;
