+++ src/main/java/com/drinkhere/drinklystore/application/service/Impl/store/RegisterStoreUseCase.java
@@ -3,8 +3,8 @@ package com.drinkhere.drinklystore.application.service.Impl.store;
 import com.drinkhere.drinklystore.clientgeocoding.dto.Coordinates;
 import com.drinkhere.drinklystore.clientgeocoding.service.GeocodingUseCase;
 import com.drinkhere.drinklystore.common.annotation.ApplicationService;
-import com.drinkhere.drinklystore.domain.dto.request.RegisterStoreRequest;
-import com.drinkhere.drinklystore.domain.dto.response.StoreResponse;
+import com.drinkhere.drinklystore.domain.dto.store.request.RegisterStoreRequest;
+import com.drinkhere.drinklystore.domain.dto.store.response.StoreResponse;
 import com.drinkhere.drinklystore.domain.entity.store.Store;
 import com.drinkhere.drinklystore.domain.service.store.StoreCommandService;
 import com.drinkhere.drinklystore.domain.service.store.StoreQueryService;
