+++ src/main/java/com/drinkhere/drinklystore/application/service/Impl/store/UpdateStoreUseCase.java
@@ -1,8 +1,8 @@
 package com.drinkhere.drinklystore.application.service.Impl.store;
 
 import com.drinkhere.drinklystore.common.annotation.ApplicationService;
-import com.drinkhere.drinklystore.domain.dto.request.UpdateStoreRequest;
-import com.drinkhere.drinklystore.domain.dto.response.GetStoreResponse;
+import com.drinkhere.drinklystore.domain.dto.store.request.UpdateStoreRequest;
+import com.drinkhere.drinklystore.domain.dto.store.response.GetStoreResponse;
 import com.drinkhere.drinklystore.domain.entity.store.Store;
 import com.drinkhere.drinklystore.domain.service.store.StoreCommandService;
 import com.drinkhere.drinklystore.domain.service.store.StoreQueryService;
