+++ src/main/java/com/drinkhere/drinklystore/application/service/UpdateImagesUseCase.java
@@ -1,7 +1,7 @@
 package com.drinkhere.drinklystore.application.service;
 
-import com.drinkhere.drinklystore.domain.dto.request.StoreImageUpdateRequest;
-import com.drinkhere.drinklystore.domain.dto.response.StoreResponse;
+import com.drinkhere.drinklystore.domain.dto.store.request.StoreImageUpdateRequest;
+import com.drinkhere.drinklystore.domain.dto.store.response.StoreResponse;
 
 public interface UpdateImagesUseCase {
     StoreResponse updateImages(Long storeId, StoreImageUpdateRequest storeImageUpdateRequest);
