+++ src/main/java/com/drinkhere/drinklystore/domain/dto/freedrink/response/GetFreeDrinkHistoriesResponseV2.java
@@ -1,8 +1,8 @@
-package com.drinkhere.drinklystore.domain.dto.response;
+package com.drinkhere.drinklystore.domain.dto.freedrink.response;
 
 import java.util.List;
 
-public record GetFreeDrinkHistoriesResponse(
+public record GetFreeDrinkHistoriesResponseV2(
         int usedCount,
         List<DrinkHistory> drinksHistory
 ) {
@@ -11,6 +11,7 @@ public record GetFreeDrinkHistoriesResponse(
     public record DrinkHistory(
             Long freeDrinkHistoryId,
             String storeName,
+            Long providedDrinkImageId,
             String providedDrink,
             String usageDate
     ) {}
