--- src/main/java/com/drinkhere/drinklystore/domain/dto/response/GetAllStoreListViewResponse.java
@@ -1,4 +1,0 @@
-package com.drinkhere.drinklystore.domain.dto.response;
-
-public record GetAllStoreListViewResponse() {
-}
