--- src/main/java/com/drinkhere/drinklystore/domain/dto/response/MemberIdResponse.java
@@ -1,6 +1,0 @@
-package com.drinkhere.drinklystore.domain.dto.response;
-
-public record MemberIdResponse(
-        Long memberId
-) {
-}
