--- src/main/java/com/drinkhere/drinklystore/domain/dto/response/OwnerIdResponse.java
@@ -1,6 +1,0 @@
-package com.drinkhere.drinklystore.domain.dto.response;
-
-public record OwnerIdResponse(
-        Long ownerId
-) {
-}
