+++ src/main/java/com/drinkhere/drinklystore/domain/dto/store/request/GetAllStoresMapViewRequest.java
@@ -1,4 +1,4 @@
-package com.drinkhere.drinklystore.domain.dto.request;
+package com.drinkhere.drinklystore.domain.dto.store.request;
 
 public record GetAllStoresMapViewRequest(
         String latitude,
