+++ src/main/java/com/drinkhere/drinklystore/domain/dto/store/request/ImageInfo.java
@@ -1,4 +1,4 @@
-package com.drinkhere.drinklystore.domain.dto.request;
+package com.drinkhere.drinklystore.domain.dto.store.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
