+++ src/main/java/com/drinkhere/drinklystore/domain/dto/store/response/GetAllStoreListViewResponse.java
@@ -1,0 +1,4 @@
+package com.drinkhere.drinklystore.domain.dto.store.response;
+
+public record GetAllStoreListViewResponse() {
+}
