+++ src/main/java/com/drinkhere/drinklystore/domain/dto/store/response/MemberIdResponse.java
@@ -1,0 +1,6 @@
+package com.drinkhere.drinklystore.domain.dto.store.response;
+
+public record MemberIdResponse(
+        Long memberId
+) {
+}
