+++ src/main/java/com/drinkhere/drinklystore/domain/dto/store/response/OwnerIdResponse.java
@@ -1,0 +1,6 @@
+package com.drinkhere.drinklystore.domain.dto.store.response;
+
+public record OwnerIdResponse(
+        Long ownerId
+) {
+}
