+++ src/main/java/com/drinkhere/drinklystore/domain/entity/store/StoreImage.java
@@ -18,7 +18,6 @@ public class StoreImage {
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-
     @JoinColumn(name = "fk_store_id", nullable = false)
     private Store store;
