+++ src/main/java/com/drinkhere/drinklystore/domain/service/store/StoreCommandService.java
@@ -1,8 +1,8 @@
 package com.drinkhere.drinklystore.domain.service.store;
 
 import com.drinkhere.drinklystore.common.annotation.DomainService;
-import com.drinkhere.drinklystore.domain.dto.request.ImageInfo;
-import com.drinkhere.drinklystore.domain.dto.request.UpdateStoreRequest;
+import com.drinkhere.drinklystore.domain.dto.store.request.ImageInfo;
+import com.drinkhere.drinklystore.domain.dto.store.request.UpdateStoreRequest;
 import com.drinkhere.drinklystore.domain.entity.store.Store;
 import com.drinkhere.drinklystore.domain.entity.store.StoreImage;
 import com.drinkhere.drinklystore.domain.enums.StoreImageType;
